+++ docker/setup-container/scripts/dockerfile-setup-container.sh
@@ -58,6 +58,10 @@ installBaseApps() {
   apt-get install -y picom
   apt-get install -y python3
   apt-get install -y python3-pyqt5
+  apt-get install -y python3-click
+  apt-get install -y python3-google-api-python-client 
+  apt-get install -y python3-google-auth-httplib2 
+  apt-get install -y python3-google-auth-oauthlib    
   apt-get install -y python3-loguru
   apt-get install -y python3-requests
   apt-get install -y python3-websocket